% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetNeighborTibble.R
\name{GetNeighborTibble}
\alias{GetNeighborTibble}
\title{Given a MultiplexObject object, creates a tibble of points and dummy points
with neighbor counts. For internal use by `SymmetricSlac` function.}
\usage{
GetNeighborTibble(obj, r, width, height, saturation, n_dummy_each)
}
\description{
Given a MultiplexObject object, creates a tibble of points and dummy points
with neighbor counts. For internal use by `SymmetricSlac` function.
}
