% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteractionSimulation.R
\name{InteractionSimulation}
\alias{InteractionSimulation}
\title{Function for simulating data with positive or negative interaction}
\usage{
InteractionSimulation(
  n_each,
  width,
  height,
  r_interaction,
  pos_m,
  neg_m,
  always_random_interaction = TRUE
)
}
\arguments{
\item{n_each}{numeric vector where each entry is the number of cells to be
simulated for a given type of point. Length of this vector determines}

\item{width}{number indicating the width of the desired window of simulation}

\item{height}{number indicating the height of the desired window of simulation}

\item{r_interaction}{number indicating the radius at which two points can be
said to be interacting}

\item{pos_m}{symmetric T x T matrix where T is the number of types of points.
Each entry should be between 0 and 1, and rows should sum to at most 1. Row i,
column j represents the probability of a point of type i *positively*
interacting with a point of type j, i.e. being generated within `r_interaction`
microns of a point of type j.}

\item{neg_m}{Defined analogously to `pos_m`, but for negative interactions.}

\item{always_random_interaction}{If `TRUE`, a point is always chosen to
randomly interact positively or negatively with another point. If `FALSE`,
the type of interaction will only be selected randomly if the point interacts
both positively and negatively with points of other types.}
}
\description{
Function for simulating data with positive or negative interaction
}
