% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeighborModelMatrix.R
\name{NeighborModelMatrix}
\alias{NeighborModelMatrix}
\title{Creates model matrix for neighbors. Designed to handle edge case where there
is only one type of point without breaking.}
\usage{
NeighborModelMatrix(neighbor_tibble)
}
\arguments{
\item{neighbor_tibble}{`neighbor_tib` from `GetNeighborTibble` function.}
}
\description{
Creates model matrix for neighbors. Designed to handle edge case where there
is only one type of point without breaking.
}
