% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SymmetricSLAC.R
\name{SymmetricSLAC}
\alias{SymmetricSLAC}
\title{Implementation of SLAC model used in simulation and application of data in
"SLAC: A Data Augmentation Approach to Modeling Multiplex Imaging Data."}
\usage{
SymmetricSLAC(
  mltplx_object,
  r,
  width,
  height,
  n_aux_each,
  fit_method = c("glm", "glmnet", "spikeslab"),
  n_iter = NULL,
  saturation = Inf
)
}
\arguments{
\item{mltplx_object}{object of class MltplxObject (from DIMPLE package).}

\item{r}{radius of interaction between different points.}

\item{width}{width of observation window of observed point process.}

\item{height}{height of observation window of observed point process.}

\item{n_aux_each}{number of auxiliary points per cell type. Can be either a
single scalar number, or a named vector with length equal to the number of
cell types in `mltplx_object`, and names corresponding to the types of cells in
`mltplx_object`.}

\item{fit_method}{method by which glm can be fit. "glm" will use the `glm`
function, "glmnet" will use `cv.glmnet`, and "spikeslab" will use a spike-and-
slab formulation implemented by the `BoomSpikeSlab` package.}

\item{n_iter}{number of iterations to run the spike-and-slab model for. Must
be provided when `fit_method` is set to "spikeslab."}

\item{saturation}{optional. This term can be used to specify a maximum number
of neighbors that a point can have, in the manner of Rajala et al. 2019. Its
usage is generally not recommended unless you know what you are doing.}
}
\description{
Implementation of SLAC model used in simulation and application of data in
"SLAC: A Data Augmentation Approach to Modeling Multiplex Imaging Data."
}
